Iri parsing performance improvement
The change uses a stateful parser implementation with mutable fields
that is aproximately 1800 times faster than previous.

With this change the difference in performance compared to akka's
uri parsing and jena's iri parsing is aproximately 5-5.5x (slower).
This is pretty acceptable considering the fact that the parser
implementation automatically normalizes the iris and supports a
wider character set.

Before:
```
[info] Benchmark                   Mode  Cnt     Score    Error  Units
[info] Parsing.parseAkkaUri       thrpt   10  1459,589 ± 57,703  ops/s
[info] Parsing.parseIri           thrpt   10     0,141 ±  0,001  ops/s
[info] Parsing.parseJenaIri       thrpt   10  1415,221 ± 18,911  ops/s
```

After:
```
[info] Benchmark              Mode  Cnt     Score    Error  Units
[info] Parsing.parseAkkaUri  thrpt   10  1383,506 ± 27,769  ops/s
[info] Parsing.parseIri      thrpt   10   263,601 ±  4,341  ops/s
[info] Parsing.parseJenaIri  thrpt   10  1450,212 ± 15,990  ops/s
```